Automated g4 rollback of commit b1c51a16cf12ee6723786d476ff4a498edc9a3a5.
*** Reason for rollback ***

Xcode 10.2.1 fixes this issue

*** Original change description ***

Force -fmodule-map-file= flags to be absolute paths to fix debugging in Swift 5.0.

Forcing the paths to be absolute seems like the opposite of what we want, and long-term it is. But lldb in Swift 5.0 does not remap relative -fmodule-map-file= paths correctly, so in order to salvage the debugging experience we have to do this temporarily.

RELNOTES: None.
PiperOrigin-RevId: 244212698